+++ src/components/SideBarContainer/useSidebar.ts
@@ -1,0 +1,11 @@
+import { useCallback, useState } from 'react';
+
+export const useSidebar = () => {
+	const [isSidebarOpened, setIsSidebarOpened] = useState(false);
+
+	const toggleSidebar = useCallback(() => {
+		setIsSidebarOpened((isOpened) => !isOpened);
+	}, []);
+
+	return { isSidebarOpened, toggleSidebar };
+};
